requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-SocketIO==5.4.1
WTForms==3.1.2
email-validator==2.2.0
python-dotenv==1.0.1
numpy==2.1.3
pandas==2.2.3
geopandas==1.0.1
XlsxWriter==3.2.0
gunicorn==23.0.0
eventlet==0.38.0
pillow==11.1.0